requirements.txt
accelerate==1.7.0
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.3.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
astroid==3.2.3
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.2.0
azure-core==1.30.1
azure-identity==1.16.0
Babel==2.14.0
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.3
bitsandbytes==0.43.3
bleach==6.1.0
build==1.2.1
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==1.0.12
ci-info==0.3.0
click==8.1.7
clickhouse-connect==0.7.19
coloredlogs==15.0.1
comm==0.2.2
configobj==5.0.8
configparser==7.1.0
contourpy==1.2.1
cryptography==42.0.5
cycler==0.12.1
dataclasses-json==0.6.4
datasets==2.18.0
debugpy==1.8.1
decorator==5.1.1
deepeval==0.21.39
defusedxml==0.7.1
Deprecated==1.2.14
dill==0.3.8
dirtyjson==1.0.8
distro==1.9.0
docx2txt==0.8
duckdb==1.1.0
EbookLib==0.18
effdet==0.4.1
einops==0.7.0
emoji==2.11.1
et-xmlfile==1.1.0
etelemetry==0.3.1
exceptiongroup==1.2.0
execnet==2.1.1
executing==2.0.1
faiss-cpu==1.8.0
fastapi==0.115.9
fastjsonschema==2.19.1
filelock==3.13.1
filetype==1.2.0
fitz==0.0.1.dev2
flash-attn==2.5.9.post1
flatbuffers==24.3.25
fonttools==4.51.0
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.2.0
ftfy==6.2.0
gekko==1.1.1
google-auth==2.29.0
googleapis-common-protos==1.63.0
greenlet==3.0.3
grpcio==1.62.2
grpcio-tools==1.62.2
h11==0.14.0
h2==4.1.0
hf-xet==1.1.2
hnswlib==0.8.0
hpack==4.0.0
httpcore==1.0.4
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.32.4
humanfriendly==10.0
hyperframe==6.0.1
idna==3.6
importlib-metadata==7.0.0
importlib_resources==6.4.0
iniconfig==2.0.0
iopath==0.1.10
ipykernel==6.29.3
ipython==8.18.1
ipywidgets==8.1.2
isodate==0.6.1
isoduration==20.11.0
isort==5.13.2
jedi==0.19.1
Jinja2==3.1.3
joblib==1.4.0
json5==0.9.24
jsonlines==4.0.0
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-lsp==2.2.4
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyter_server==2.13.0
jupyter_server_terminals==0.5.3
jupyterlab==4.1.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.4
jupyterlab_widgets==3.0.10
kiwisolver==1.4.5
kubernetes==29.0.0
langchain==0.3.25
langchain-chroma==0.1.0
langchain-community==0.3.24
langchain-core==0.3.63
langchain-experimental==0.0.57
langchain-huggingface==0.2.0
langchain-nomic==0.1.4
langchain-openai==0.0.5
langchain-text-splitters==0.3.8
langdetect==1.0.9
langgraph==0.0.37
langsmith==0.1.147
layoutparser==0.3.4
loguru==0.7.2
looseversion==1.3.0
lxml==5.2.1
lz4==4.3.3
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
matplotlib==3.8.4
matplotlib-inline==0.1.6
matplotlib-venn==0.11.10
mccabe==0.7.0
mdurl==0.1.2
mistune==3.0.2
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
msal==1.28.0
msal-extensions==1.1.0
msg-parser==1.2.0
multidict==6.0.5
-e git+ssh://git@github.com/riedlerm/multimodal_rag_for_industry.git@2dbfed822d8c7c86d4819ea720dc00cd31d3b409#egg=multimodal_rag_setup_file
multiprocess==0.70.16
mypy-extensions==1.0.0
nbclient==0.10.0
nbconvert==7.16.3
nbformat==5.10.3
nest-asyncio==1.6.0
networkx==3.2.1
nibabel==5.2.1
ninja==1.11.1.1
nipype==1.8.6
nltk==3.8.1
nomic==3.5.1
notebook==7.1.2
notebook_shim==0.2.4
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oauthlib==3.2.2
olefile==0.47
omegaconf==2.3.0
onnx==1.14.1
onnxruntime==1.16.3
open-clip-torch==2.24.0
openai==1.23.2
opencv-python==4.9.0.80
openpyxl==3.1.2
opentelemetry-api==1.24.0
opentelemetry-exporter-otlp-proto-common==1.24.0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
opentelemetry-instrumentation==0.45b0
opentelemetry-instrumentation-asgi==0.45b0
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-proto==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-semantic-conventions==0.45b0
opentelemetry-util-http==0.45b0
optimum==1.25.3
orjson==3.10.1
overrides==7.7.0
packaging==23.2
pandas==2.2.1
pandocfilters==1.5.1
parso==0.8.3
pathlib==1.0.1
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.0
peft==0.11.1
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.2.0
pluggy==1.5.0
portalocker==2.8.2
posthog==3.5.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
protobuf==3.20.3
prov==2.0.1
psutil==5.9.8
ptyprocess==0.7.0
pulsar-client==3.5.0
pure-eval==0.2.2
pyarrow==15.0.2
pyarrow-hotfix==0.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycocotools==2.0.7
pycparser==2.21
pycryptodome==3.20.0
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
pydot==3.0.1
Pygments==2.17.2
PyJWT==2.8.0
pylint==3.2.5
PyMuPDF==1.25.5
pypandoc==1.13
pyparsing==3.1.2
pypdf==4.2.0
PyPDF2==3.0.1
pypdfium2==4.29.0
PyPika==0.48.9
pyproject_hooks==1.0.0
pysbd==0.3.4
pysqlite3==0.5.3
pytesseract==0.3.10
pytest==8.2.0
pytest-repeat==0.9.3
pytest-xdist==3.6.1
python-dateutil==2.9.0.post0
python-docx==1.1.0
python-dotenv==1.0.1
python-iso639==2024.2.7
python-json-logger==2.0.7
python-magic==0.4.27
python-multipart==0.0.9
python-pptx==0.6.21
pytz==2024.1
pyxnat==1.6.2
PyYAML==6.0.1
pyzmq==25.1.2
qdrant-client==1.9.0
radar-chart==0.0.5
ragas==0.1.7
rapidfuzz==3.8.1
rdflib==6.3.2
referencing==0.34.0
regex==2023.12.25
requests==2.31.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rouge==1.0.1
rpds-py==0.18.0
rsa==4.9
safetensors==0.4.3
scikit-learn==1.4.2
scipy==1.13.0
seaborn==0.13.2
Send2Trash==1.8.2
sentence-transformers==4.1.0
sentencepiece==0.1.99
sentry-sdk==2.1.1
shellingham==1.5.4
simplejson==3.19.3
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.29
stack-data==0.6.3
starlette==0.45.3
striprtf==0.0.26
sympy==1.13.1
tabulate==0.9.0
tenacity==8.2.3
terminado==0.18.1
threadpoolctl==3.5.0
tiktoken==0.5.2
timm==0.9.16
tinycss2==1.2.1
tokenizers==0.21.1
tomli==2.0.1
tomlkit==0.13.0
torch==2.6.0
torchaudio==2.5.1+cu121
torchvision==0.21.0
tornado==6.4
tqdm==4.66.2
traitlets==5.14.2
traits==6.3.2
transformers==4.52.4
transformers-stream-generator==0.0.5
triton==3.2.0
typer==0.12.3
types-python-dateutil==2.9.0.20240316
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2024.1
umap==0.1.1
unstructured==0.10.19
unstructured-inference==0.6.6
unstructured.pytesseract==0.3.12
uri-template==1.3.0
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
websockets==12.0
widgetsnbextension==4.0.10
wrapt==1.16.0
xformers==0.0.26.post1
xlrd==2.0.1
XlsxWriter==3.2.0
xxhash==3.4.1
yarl==1.9.4
zipp==3.18.1
zstandard==0.23.0
